fix: chnage js build path
